update setup, create result/ and other fix

+++ space/setup/index.php
@@ -27,22 +27,27 @@
         }
     }
     
+    // removing database, files and session
     if (isset($_POST["action"])){
         if ($_POST["action"] == "RESET"){
-            $dir = "../" . $spacedir;
-            $it = new RecursiveDirectoryIterator($dir, RecursiveDirectoryIterator::SKIP_DOTS);
-            $files = new RecursiveIteratorIterator($it, RecursiveIteratorIterator::CHILD_FIRST);
-            foreach ($files as $file){
-                if ($file->isDir()){
-                    rmdir($file->getRealPath());
-                }else{
-                    unlink($file->getRealPath());
-                }
-            }
-            rmdir($dir);
             for ($i = 0; $i < count($tables); $i++){
                 mysqli_query($connection, "DROP TABLE " . $tables[$i]);
             }
+
+            if (isset($_POST['all'])){
+                $dir = "../" . $spacedir;
+                $it = new RecursiveDirectoryIterator($dir, RecursiveDirectoryIterator::SKIP_DOTS);
+                $files = new RecursiveIteratorIterator($it, RecursiveIteratorIterator::CHILD_FIRST);
+                foreach ($files as $file){
+                    if ($file->isDir()){
+                        rmdir($file->getRealPath());
+                    }else{
+                        unlink($file->getRealPath());
+                    }
+                }
+                rmdir($dir);
+            }
+            
             session_unset();
             session_destroy();
             header("Location: ../setup/");
@@ -50,20 +55,35 @@
         }
     }
 
-    
-    if (!is_dir('../'.$spacedir)) mkdir('../'.$spacedir,0777,true);
+    // creating directory and database
+    if (!is_dir('../'.$spacedir)) {
+        mkdir('../'.$spacedir,0777,true);
+        echo "Directory create sucessfuly!<br><br>";
+    }
 
     if (!in_array("blackhole", $tables)){
         $query = "CREATE TABLE blackhole (
             id INT(10) PRIMARY KEY, 
+            owner_ VARCHAR(20) NOT NULL,
             codename_ VARCHAR(20) NOT NULL, 
             sector_ VARCHAR(2) NOT NULL, 
-            path_ VARCHAR(71) NOT NULL,
             name_ VARCHAR(50) NOT NULL,
-            limit_ INT(4) NOT NULL,
-            owner_ VARCHAR(20) NOT NULL
+            savedname_ VARCHAR(71) NOT NULL,
+            size_ VARCHAR(11) NOT NULL,
+            hours_ INT(4) NOT NULL
         )";
-        if (mysqli_query($connection, $query)) echo "Blackhole create successfully!<br>";
+        if (mysqli_query($connection, $query)) echo "Blackhole create sucessfully!<br><br>";
+    }
+
+    if (!in_array("logging", $tables)){
+        $query = "CREATE TABLE logging (
+            id INT(10) PRIMARY KEY,
+            owner_ VARCHAR(20) NOT NULL,
+            codename_ VARCHAR(20) NOT NULL,
+            name_ VARCHAR(255) NOT NULL,
+            size_ VARCHAR(11) NOT NULL
+        )";
+        if (mysqli_query($connection, $query)) echo "Logline create sucessfully!<br><br>";
     }
 
     if (!in_array("traceline", $tables)){
@@ -72,7 +92,7 @@
             penname_ VARCHAR(20) NOT NULL,
             message_ VARCHAR(1023) NOT NULL
         )";
-        if (mysqli_query($connection, $query)) echo "Traceline create successfully!<br>";
+        if (mysqli_query($connection, $query)) echo "Traceline create sucessfully!<br><br>";
     }
 
     if (!in_array("spaceship", $tables)){
@@ -82,22 +102,13 @@
             password_ VARCHAR(255) NOT NULL,
             level_ INT(2) NOT NULL
         )";
-        if (mysqli_query($connection, $query)) echo "Spaceship create successfully!<br>";
+        if (mysqli_query($connection, $query)) echo "Spaceship create sucessfully!<br><br>";
         
         $time = time();
         $password = password_hash($defaultpass, PASSWORD_DEFAULT);
         mysqli_query($connection, "INSERT INTO spaceship (id, username_, password_, level_) VALUES ('$time', '$defaultuser', '$password', 1)");
     }
     
-    if (!in_array("logging", $tables)){
-        $query = "CREATE TABLE logging (
-            id INT(10) PRIMARY KEY,
-            codename_ VARCHAR(20) NOT NULL,
-            name_ VARCHAR(255) NOT NULL,
-            size_ varchar(10) NOT NULL
-        )";
-        if (mysqli_query($connection, $query)) echo "Logline create successfully!<br>";
-    }
 
     if (empty($tablelist)){
         $tablelist = querying("SHOW TABLES");
@@ -109,7 +120,6 @@
             }
         }
     }
-
 ?>
 
 <!DOCTYPE html>
@@ -122,42 +132,42 @@
     <link rel="stylesheet" href="../styles.css">
 
     <style>
-        .menu{display: flex;}
-        .item{margin: 10px;}
-        .item input:hover{cursor: pointer;}
+        .menu input[type="submit"]{width: 100px;}
+        .menu td{padding: 5px 2px;}
     </style>
 </head>
 
 <body>    
-    <div class="menu">
-        <form class="item" action="" method="post"> 
-            <input type="submit" name="action" value="RESET">
-        </form>
-        <a class="item" href="../"><input type="submit" value="BLACKHOLE"></a>
-    </div>
-    <br><hr>
+    <table class="menu">
+        <tr>
+            <form action="" method="post"> 
+                <td><input type="submit" name="action" value="RESET"></td>
+                <td><input type="checkbox" name="all"> Delete Files</td>
+            </form>
+        </tr>
+        <tr><td><a href="../"><input type="submit" value="BLACKHOLE"></a></td></tr>
+    </table>
+    <br>
 
-    <?php for ($i = 0; $i < count($tables); $i++): ?>
+    <?php for ($i = 0; $i < count($tables); $i++): ?><hr>
         <h3><?=$tables[$i]?></h3>
         <div style="display: flex;">
             <?php for ($ii = 0; $ii < count($headers[$i]); $ii++): ?>
                 <div style="margin: 0 10px;"><?=$headers[$i][$ii]?></div>
             <?php endfor ?>
-        </div><br><hr>
+        </div><br>
     <?php endfor ?>
-    
-    <br><br><br><br><br>
 
-    <div class="footer"><br>
+    <br><br><br><br><br><br><br><br>
+
+    <div class="footer">
         <script async defer src="https://buttons.github.io/buttons.js"></script>
         <table>
-            <tr>
-                <td style="padding: 0 10px;"><a href="https://buymeacoffee.com/vanila" target="_blank"><img src="https://cdn.buymeacoffee.com/buttons/default-orange.png" alt="Buy Me A Coffee" height="28" width="146"></a></td>
-                <td style="padding: 0 10px;"><a class="github-button" href="https://github.com/saandhikaa" data-size="large" aria-label="Follow @saandhikaa on GitHub">Follow @saandhikaa</a>            </td>
-            </tr>
+            <tr><td><a href="https://buymeacoffee.com/vanila" target="_blank"><img src="https://cdn.buymeacoffee.com/buttons/default-orange.png" alt="Buy Me A Coffee" height="41" width="174"></a></td></tr>
+            <tr><td><a class="github-button" href="https://github.com/saandhikaa" data-size="large" aria-label="Follow @saandhikaa on GitHub">Follow @saandhikaa</a></td></tr>
         </table>
-                    
-        <p class="copyright">Copyright 2023 - <?= $_SERVER['SERVER_NAME'] ?> - All Rights Reserved</p>
+
+        <p class="copyright">Copyright 2023 - <?= $server ?> - All Rights Reserved</p>
     </div>
 </body>
 </html>
